+++ apps/extension/src/app/common/transactions/stacks/use-stacks-broadcast-transaction.tsx
@@ -1,29 +1,22 @@
 import { useMemo, useState } from 'react';
 import { useNavigate } from 'react-router';
 
-import { AuthType, StacksTransactionWire } from '@stacks/transactions';
+import { StacksTransactionWire } from '@stacks/transactions';
 
 import type { CryptoCurrency } from '@leather.io/models';
-import { delay, isError, isString } from '@leather.io/utils';
+import { isError, isString } from '@leather.io/utils';
 
-import { finalizeTxSignature } from '@shared/actions/finalize-tx-signature';
 import { logger } from '@shared/logger';
 import { RouteUrls } from '@shared/route-urls';
 
-import { useDefaultRequestParams } from '@app/common/hooks/use-default-request-search-params';
 import { useSubmitTransactionCallback } from '@app/common/hooks/use-submit-stx-transaction';
 import {
   StacksTransactionActionType,
   stacksTransactionToHex,
 } from '@app/common/transactions/stacks/transaction.utils';
 import { useToast } from '@app/features/toasts/use-toast';
-import { useTransactionRequest } from '@app/store/transactions/requests.hooks';
 import { useSignStacksTransaction } from '@app/store/transactions/transaction.hooks';
 import { LoadingKeys } from '@app/store/ui/ui.hooks';
-
-async function simulateShortDelayToAvoidUndefinedTabId() {
-  await delay(1000);
-}
 
 interface UseStacksBroadcastTransactionArgs {
   actionType?: StacksTransactionActionType;
@@ -40,17 +33,14 @@ export function useStacksBroadcastTransaction({
 }: UseStacksBroadcastTransactionArgs) {
   const signStacksTransaction = useSignStacksTransaction();
   const [isBroadcasting, setIsBroadcasting] = useState(false);
-  const { frameId, tabId } = useDefaultRequestParams();
-  const requestToken = useTransactionRequest();
 
   const navigate = useNavigate();
   const toast = useToast();
 
   const isCancelTransaction = actionType === StacksTransactionActionType.Cancel;
   const isIncreaseFeeTransaction = actionType === StacksTransactionActionType.IncreaseFee;
-  const isRpcRequest = actionType === StacksTransactionActionType.RpcRequest;
 
-  const showSummaryPage = !isCancelTransaction && !isIncreaseFeeTransaction && !isRpcRequest;
+  const showSummaryPage = !isCancelTransaction && !isIncreaseFeeTransaction;
 
   const broadcastTransactionFn = useSubmitTransactionCallback({
     loadingKey: LoadingKeys.SUBMIT_STACKS_TRANSACTION,
@@ -58,17 +48,6 @@ export function useStacksBroadcastTransaction({
 
   return useMemo(() => {
     function handlePreviewSuccess(signedTx: StacksTransactionWire, txid?: string) {
-      if (requestToken && tabId) {
-        finalizeTxSignature({
-          frameId,
-          requestPayload: requestToken,
-          tabId,
-          data: {
-            txRaw: stacksTransactionToHex(signedTx),
-            txId: txid,
-          },
-        });
-      }
       if (txid && redirectToSuccessPage) {
         void navigate(
           RouteUrls.SentStxTxSummary.replace(':symbol', token.toLowerCase()).replace(
@@ -88,26 +67,20 @@ export function useStacksBroadcastTransaction({
       }
       try {
         setIsBroadcasting(true);
-        const isSponsored = signedTx.auth?.authType === AuthType.Sponsored;
-        if (isSponsored) {
-          await simulateShortDelayToAvoidUndefinedTabId();
-          handlePreviewSuccess(signedTx);
-        } else {
-          return await broadcastTransactionFn({
-            onError(e: Error | string) {
-              const message = isString(e) ? e : e.message;
-              return navigate(RouteUrls.BroadcastError, { state: { message } });
-            },
-            onSuccess(txId) {
-              if (showSummaryPage) return handlePreviewSuccess(signedTx, txId);
-              void navigate(RouteUrls.Activity);
-              if (isCancelTransaction) return toast.success('Transaction cancelled successfully');
-              if (isIncreaseFeeTransaction) return toast.success('Fee increased successfully');
-              return;
-            },
-            replaceByFee: false,
-          })(signedTx);
-        }
+        return await broadcastTransactionFn({
+          onError(e: Error | string) {
+            const message = isString(e) ? e : e.message;
+            return navigate(RouteUrls.BroadcastError, { state: { message } });
+          },
+          onSuccess(txId) {
+            if (showSummaryPage) return handlePreviewSuccess(signedTx, txId);
+            void navigate(RouteUrls.Activity);
+            if (isCancelTransaction) return toast.success('Transaction cancelled successfully');
+            if (isIncreaseFeeTransaction) return toast.success('Fee increased successfully');
+            return;
+          },
+          replaceByFee: false,
+        })(signedTx);
       } catch (e) {
         return navigate(RouteUrls.BroadcastError, {
           state: { message: isError(e) ? e.message : 'Unknown error' },
@@ -135,9 +108,6 @@ export function useStacksBroadcastTransaction({
     };
   }, [
     isBroadcasting,
-    requestToken,
-    frameId,
-    tabId,
     redirectToSuccessPage,
     navigate,
     token,