+++ packages/provider/src/mobile.ts
@@ -10,7 +10,6 @@ initInpageProvider({
     (window as any).ReactNativeWebView.postMessage(JSON.stringify(rpcRequest));
   },
   env: {
-    platform: 'mobile',
     branch: 'replace_branch',
     commitSha: 'replace_commit_sha',
     version: 'replace_version',