+++ packages/crypto/CHANGELOG.md
@@ -492,6 +492,16 @@
   * dependencies
     * @leather.io/utils bumped to 0.46.4
 
+## [1.13.5](https://github.com/leather-io/mono/compare/@leather.io/crypto-v1.13.4...@leather.io/crypto-v1.13.5) (2026-08-10)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @leather.io/constants bumped to 0.39.0
+    * @leather.io/utils bumped to 0.52.5
+
 ## [1.13.4](https://github.com/leather-io/mono/compare/@leather.io/crypto-v1.13.3...@leather.io/crypto-v1.13.4) (2026-08-05)